# src/main/resources/order-openapi.yaml
openapi: 3.0.1
info:
  title: Order API
  description: API for managing a single order
  version: 1.0.0
servers:
  - url: https://orderservice.example.com/api

paths:
  /order:
    get:
      summary: Get the current order
      operationId: getOrder
      tags:
        - Order
      responses:
        '200':
          description: The current order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Order not found
        '500':
          description: Internal server error

    post:
      summary: Create a new order
      operationId: createOrder
      tags:
        - Order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
      responses:
        '201':
          description: Order created successfully
        '400':
          description: Invalid order data
        '500':
          description: Internal server error

    put:
      summary: Update the current order
      operationId: updateOrder
      tags:
        - Order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
      responses:
        '200':
          description: Order updated successfully
        '400':
          description: Invalid order data
        '404':
          description: Order not found
        '500':
          description: Internal server error

components:
  schemas:
    Order:
      type: object
      properties:
        orderId:
          type: string
          description: Unique identifier for the order
        customerName:
          type: string
          description: The name of the customer who placed the order
        orderDate:
          type: string
          format: date-time
          description: The date when the order was placed
        totalAmount:
          type: number
          format: float
          description: The total amount of the order
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'

    OrderItem:
      type: object
      properties:
        productId:
          type: string
          description: ID of the product in the order
        quantity:
          type: integer
          description: Quantity of the product ordered
        price:
          type: number
          format: float
          description: Price of a single unit of the product